Implement filter query language
Filter is now a comma-separated list of patterns to search for. A task is shown if it matches any of the patterns.
Each pattern consists of a space-separated list of strings, possibly preceded by ‘!’.

E.g., the query “array !sorted, btree3” shows “Array.sort” and “BTree3.insert”, but not “SortedArray.insert”.